feat : implement AuthRequestHeaderValidator
